Created universal projects and platform specific mkspecs
Now CoreLibrary pro files for lib, DLL (and SO) and unit test projects are no longer separate per platform. There is a single pro file for lib, one for DLL/SO and one for unit test. The projects support linux, windows 32 bit and windows 64bit. Linux is assumed to always be building for single platform e.g. usually always 64bit. The windows settings expect the user to be running MSVC 2015 or greater.