Preparations for V1.4.0.0
More work toclarify and tidy the doxygen comments within the code. Also implemented a specialized cell calss called CellDouble for a specialized version of the CSVGrid when all that is needed is double precision data. In this case all data is stored as a double and not as a std::string like the general usage case.